requirements.txt
openai>=1.66.0
flask>=3.1.0
python-dotenv>=1.0.1
